Add Python 3.14 support and update CI/CD workflows (#440)
* Add Python 3.14 support and update CI/CD workflows
   - Add Python 3.14 to supported versions in pyproject.toml
   - Update GitHub Actions workflows to test Python 3.14
   - Update CI matrix to exclude Python 3.12 on macOS/Windows for efficiency
   - Upgrade GitHub Actions to latest versions (checkout@v5, setup-python@v6)
   - Switch from conda to pip for deploy workflow dependencies
   - Update pytest to use auto parallelization instead of fixed cores
   - update CHANGES
* allow numpy >=2.3.3, exclude 2.3.0-2.3.2
   Initial evidence with pymbar (PR choderalab/pymbar#566) indicates that at least
   release 2.3.3 of numpy produces stable results again.

   Thus, exclude 2.3.0, 2.3.1, 2.3.2 (we did not explicitly test 2.3.[12] but
   exclude them because we known that 2.3.3 works).

* explicit array typing ti_gaussian_quadrature_.py to make mypy happy

+++ src/alchemlyb/estimators/ti_gaussian_quadrature_.py
@@ -386,7 +386,7 @@ class TI_GQ(BaseEstimator, _EstimatorMixOut):
         d_deltas_squared = np.insert(d_deltas_squared, 0, [0.0], axis=0)
         d_deltas_squared = np.append(d_deltas_squared, [0.0], axis=0)
         # build matrix of deltas between each state
-        adelta = np.zeros((len(deltas), len(deltas)))
+        adelta = np.zeros((len(deltas), len(deltas)), dtype=float)
         ad_delta = np.zeros_like(adelta)
 
         for j in range(len(deltas)):
@@ -403,7 +403,7 @@ class TI_GQ(BaseEstimator, _EstimatorMixOut):
             adelta += np.diagflat(np.array(out), k=j)
             ad_delta += np.diagflat(np.array(dout), k=j)
 
-        adelta = adelta - adelta.T  # type: ignore[assignment]
+        adelta = adelta - adelta.T
         ad_delta = (ad_delta + ad_delta.T) - 2 * np.diagflat(d_deltas_squared)
         # yield standard delta_f_ cumulative free energies from one state to another
         self._delta_f_ = pd.DataFrame(adelta, columns=index_list, index=index_list)
